LayoutTests/inspector/debugger/break-on-uncaught-exception-expected.txt

CONSOLE MESSAGE: line 7: throwing TestError
CONSOLE MESSAGE: line 8: TestError
CONSOLE MESSAGE: line 22: testFinally
CONSOLE MESSAGE: line 7: throwing TestError
CONSOLE MESSAGE: line 26: finally
CONSOLE MESSAGE: line 8: TestError
CONSOLE MESSAGE: line 31: testThrowingThruNativeCode
CONSOLE MESSAGE: line 7: throwing TestError
CONSOLE MESSAGE: line 8: TestError
CONSOLE MESSAGE: line 12: testCatch
CONSOLE MESSAGE: line 7: throwing TestError
CONSOLE MESSAGE: line 16: catch TestError
CONSOLE MESSAGE: line 18: DONE
Checking pause locations when pausing on uncaught exceptions.


== Running test suite: BreakOnUncaughtException
-- Running test case: BreakOnUncaughtException.UncaughtExceptionBasic
PASS: Should pause for exception.
PAUSE AT doThrow:8:22
Uncaught exception in test page: TestError [break-on-exception-tests.js:8]

-- Running test case: BreakOnUncaughtException.UncaughtExceptionInFinally
PASS: Should pause for exception.
PAUSE AT doThrow:8:22
Uncaught exception in test page: TestError [break-on-exception-tests.js:8]

-- Running test case: BreakOnUncaughtException.UncaughtExceptionThruNativeCode
PASS: Should pause for exception.
PAUSE AT doThrow:8:22
Uncaught exception in test page: TestError [break-on-exception-tests.js:8]

-- Running test case: BreakOnUncaughtException.CaughtException
PASS: Should not pause, exception was caught.
